test/fixtures/wpt/wasm/webapi/esm-integration/resolve-export.tentative.html

<!DOCTYPE html>
<title>Check ResolveExport on invalid re-export from WebAssembly</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    const test_load = async_test(
        "Re-export of missing Wasm export should result in SyntaxError.");

    window.log = [];
    window.addEventListener("error", ev => {
      test_load.step(() => assert_equals(ev.error.constructor, SyntaxError));
      log.push(ev.message);
    });

    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(log[1], 1);
    }));

    function unreachable() { log.push("unexpected"); }
</script>
<script type="module" src="./resolve-export.js"
    onerror="unreachable()" onload="log.push(1)"></script>
