third_party/blink/web_tests/external/wpt/wasm/webapi/esm-integration/resolve-export.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Re-export of missing Wasm export should result in SyntaxError.
  assert_equals: expected (number) 1 but got (undefined) undefined
Harness: the test ran to completion.
